uday/requirements.txt
livekit-agents[tavus]~=1.0
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
pydantic-ai-slim[openai,mcp]
tzdata
flask[async]
flask-cors
